requirements.txt
websocket-client
kafka-python
cassandra-driver
spark
pyspark